MVP-4112-patch: refactor history list
