+++ packages/notifi-dapp-example/next.config.mjs
@@ -1,5 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+  // output: 'export', /** BUILD ONLY */
   images: {
     formats: ['image/webp'],
   },
