+++ packages/notifi-dapp-example/src/components/HistoryList.tsx
@@ -1,12 +1,9 @@
 'use client';
 
-import { useGlobalStateContext } from '@/context/GlobalStateContext';
-import { useNotifiCardContext } from '@/context/notifi/NotifiCardContext';
-import { validateEventDetails } from '@/utils/notifiHistoryUtils';
+import { useNotifiHistory } from '@/hooks/useNotifiHistory';
 import { Types } from '@notifi-network/notifi-graphql';
-import { useNotifiClientContext } from '@notifi-network/notifi-react-card';
 import Image from 'next/image';
-import { Dispatch, SetStateAction, useEffect, useRef, useState } from 'react';
+import { Dispatch, SetStateAction } from 'react';
 
 import { HistoryListRow } from './HistoryListRow';
 import { LoadingSpinner } from './LoadingSpinner';
@@ -18,84 +15,18 @@ type HistoryListProps = {
   historyDetailEntry: Types.FusionNotificationHistoryEntryFragmentFragment | null;
 };
 
-type CursorInfo = Readonly<{
-  hasNextPage: boolean;
-  endCursor?: string | undefined;
-}>;
-
-const messagePerPage = 50;
-
 export const HistoryList: React.FC<HistoryListProps> = ({
   setHistoryDetailEntry,
   historyDetailEntry,
 }) => {
-  const { frontendClient } = useNotifiClientContext();
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const { setGlobalError } = useGlobalStateContext();
-  const [nodes, setNodes] = useState<
-    Types.FusionNotificationHistoryEntryFragmentFragment[]
-  >([]);
-  const [unreadCount, setUnreadCount] = useState<number | null>(null);
-  const [cursorInfo, setCursorInfo] = useState<CursorInfo>({
-    hasNextPage: false,
-    endCursor: undefined,
-  });
-  const historyLoaded = useRef<boolean>(false);
-  const { cardConfig } = useNotifiCardContext();
-
-  const cardEventTypeNames = new Set(
-    cardConfig?.eventTypes?.map((event) => event.name) ?? [],
-  );
-
-  // TODO: move to useNotifiHistory hook
-  const getNotificationHistory = (initialLoad?: boolean) => {
-    if (!initialLoad && !cursorInfo.hasNextPage) {
-      setGlobalError('No more notification history to fetch');
-      return;
-    }
-    if (isLoading) {
-      return;
-    }
-    setIsLoading(true);
-    frontendClient
-      .getFusionNotificationHistory({
-        first: messagePerPage,
-        after: cursorInfo.endCursor,
-        includeHidden: false,
-      })
-      .then((result) => {
-        if (!result || result.nodes?.length === 0 || !result.nodes) {
-          return;
-        }
-        // Filter out events that are not supported by the card
-        result.nodes.filter((node) => {
-          if (!node.detail || !validateEventDetails(node.detail)) return false;
-          return cardEventTypeNames.has(node.detail.sourceName);
-        });
-
-        setNodes((existing) =>
-          initialLoad
-            ? result.nodes ?? []
-            : [...existing, ...(result.nodes ?? [])],
-        );
-        setCursorInfo(result.pageInfo);
-        historyLoaded.current = true;
-      })
-      .catch((e) => {
-        setGlobalError(
-          'ERROR: Failed to fetch notification history, check console for more details',
-        );
-        console.error('Failed to fetch notification history', e);
-      })
-      .finally(() => setIsLoading(false));
-  };
-
-  useEffect(() => {
-    getNotificationHistory(true);
-    frontendClient.getUnreadNotificationHistoryCount().then(({ count }) => {
-      setUnreadCount(count);
-    });
-  }, [frontendClient]);
+  const {
+    isLoading,
+    cursorInfo,
+    nodes,
+    unreadCount,
+    setUnreadCount,
+    getNotificationHistory,
+  } = useNotifiHistory(true);
 
   // useEffect(() => {
   //   // TODO: TBD feature
@@ -148,14 +79,14 @@ export const HistoryList: React.FC<HistoryListProps> = ({
             setUnreadCount={setUnreadCount}
           />
         ))}
-      </div>
-      <div
-        className={`m-auto mt-4 text-lg font-semibold cursor-pointer ${
-          !cursorInfo.hasNextPage ? 'hidden' : ''
-        }`}
-        onClick={() => getNotificationHistory()}
-      >
-        Load more
+        <div
+          className={` flex justify-center my-4 text-lg font-semibold cursor-pointer ${
+            !cursorInfo.hasNextPage ? 'hidden' : ''
+          }`}
+          onClick={() => getNotificationHistory()}
+        >
+          <div>Load more</div>
+        </div>
       </div>
     </div>
   );
