+++ packages/notifi-dapp-example/src/hooks/useNotifiHistory.ts
@@ -1,13 +1,43 @@
+import { useGlobalStateContext } from '@/context/GlobalStateContext';
+import { useNotifiCardContext } from '@/context/notifi/NotifiCardContext';
+import { validateEventDetails } from '@/utils/notifiHistoryUtils';
+import { Types } from '@notifi-network/notifi-graphql';
 import { useNotifiClientContext } from '@notifi-network/notifi-react-card';
-import { useCallback } from 'react';
+import { useCallback, useEffect, useRef, useState } from 'react';
 
-export const useNotifiHistory = () => {
+type CursorInfo = Readonly<{
+  hasNextPage: boolean;
+  endCursor?: string | undefined;
+}>;
+
+const messagePerPage = 50;
+
+/**
+ * @param autoFetchHistoryAndUnreadCount - If true, fetches the notification history & unread count on mount
+ */
+export const useNotifiHistory = (autoFetchHistoryAndUnreadCount?: boolean) => {
+  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const { popGlobalInfoModal } = useGlobalStateContext();
   const { frontendClient } = useNotifiClientContext();
+  const [cursorInfo, setCursorInfo] = useState<CursorInfo>({
+    hasNextPage: false,
+    endCursor: undefined,
+  });
+  const [unreadCount, setUnreadCount] = useState<number | null>(null);
+  const [nodes, setNodes] = useState<
+    Types.FusionNotificationHistoryEntryFragmentFragment[]
+  >([]);
+  const { cardConfig } = useNotifiCardContext();
+  const historyLoaded = useRef<boolean>(false);
+
+  const cardEventTypeNames = new Set(
+    cardConfig?.eventTypes?.map((event) => event.name) ?? [],
+  );
 
   const markNotifiHistoryAsRead = useCallback(
     async (ids?: string[]) => {
       if (!frontendClient) return;
-
+      // TODO: implement loading state
       if (ids?.length === 0 || !ids) {
         // Mark all as read
         const res = await frontendClient.getFusionNotificationHistory({
@@ -24,12 +54,79 @@ export const useNotifiHistory = () => {
       }
 
       // Mark selected as read
-      frontendClient.markFusionNotificationHistoryAsRead({
+      await frontendClient.markFusionNotificationHistoryAsRead({
         ids,
       });
     },
     [frontendClient],
   );
 
-  return { markNotifiHistoryAsRead };
+  const getNotificationHistory = (initialLoad?: boolean) => {
+    if (!initialLoad && !cursorInfo.hasNextPage) {
+      popGlobalInfoModal({
+        message: 'No more notification history to fetch',
+        iconOrEmoji: { type: 'icon', id: 'warning' },
+        timeout: 5000,
+      });
+      return;
+    }
+    if (isLoading) {
+      return;
+    }
+
+    setIsLoading(true);
+    frontendClient
+      .getFusionNotificationHistory({
+        first: messagePerPage,
+        after: cursorInfo.endCursor,
+        includeHidden: false,
+      })
+      .then((result) => {
+        if (!result || result.nodes?.length === 0 || !result.nodes) {
+          return;
+        }
+        // Filter out events that are not supported by the card
+        result.nodes.filter((node) => {
+          if (!node.detail || !validateEventDetails(node.detail)) return false;
+          return cardEventTypeNames.has(node.detail.sourceName);
+        });
+
+        setNodes((existing) =>
+          initialLoad
+            ? result.nodes ?? []
+            : [...existing, ...(result.nodes ?? [])],
+        );
+        setCursorInfo(result.pageInfo);
+        historyLoaded.current = true;
+      })
+      .catch((e) => {
+        popGlobalInfoModal({
+          message:
+            'ERROR: Failed to fetch notification history, check console for more details',
+          iconOrEmoji: { type: 'icon', id: 'warning' },
+          timeout: 5000,
+        });
+        console.error('Failed to fetch notification history', e);
+      })
+      .finally(() => setIsLoading(false));
+  };
+
+  if (autoFetchHistoryAndUnreadCount) {
+    useEffect(() => {
+      getNotificationHistory(true);
+      frontendClient.getUnreadNotificationHistoryCount().then(({ count }) => {
+        setUnreadCount(count);
+      });
+    }, [frontendClient]);
+  }
+
+  return {
+    markNotifiHistoryAsRead,
+    isLoading,
+    nodes,
+    getNotificationHistory,
+    unreadCount,
+    cursorInfo,
+    setUnreadCount,
+  };
 };
